% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diccionario_vars_adicional.R
\docType{data}
\name{diccionario_vars_adicional}
\alias{diccionario_vars_adicional}
\title{Variables adicionales que pueden ser útiles para analizar los datos.}
\format{
Un data frame con 11 filas y 4 variables.
\describe{
\item{nombre_variable}{nombre de la variable}
\item{tipo}{tipo de la variable: caracter, numérico, lógico}
\item{detalle_variable}{explicación detallada de la variable}
\item{categorias_variable}{valores posibles de la variable}
}
}
\source{
Proyecto conjunto JEP-CEV-HRDAG.
}
\usage{
data(diccionario_vars_adicional)
}
\description{
Variables adicionales que pueden ser útiles para analizar los datos.
}
\keyword{datasets}
