% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estratificacion.R
\docType{data}
\name{estratificacion}
\alias{estratificacion}
\title{Datos que documentan las estratificaciones necesarias para replicar
los resultados del informe metodológico del proyecto conjunto CEV-HRDAG-JEP
(versión en español).}
\format{
Un data frame con 31 filas y 4 variables.
\describe{
\item{violacion}{el hecho de violencia al analizar}
\item{estimacion}{el tipo de análisis que utiliza la estratificación (p.ej., patrones de violencia por año, sexo, etc.)}
\item{estratificacion}{las variables utilizadas para estratificar las estimaciones}
\item{notas}{notas adicionales sobre la estratificación; NA si no hay notas}
}
}
\source{
Proyecto conjunto JEP-CEV-HRDAG.
}
\usage{
data(estratificacion)
}
\description{
Datos que documentan las estratificaciones necesarias para replicar
los resultados del informe metodológico del proyecto conjunto CEV-HRDAG-JEP
(versión en español).
}
\keyword{datasets}
