% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_standard_cev.R
\name{filter_standard_cev}
\alias{filter_standard_cev}
\title{Filter records to replicate results presented in the CEV methodology report.}
\usage{
filter_standard_cev(replicates, violation, perp_change = TRUE)
}
\arguments{
\item{replicates}{Dataframe with data from all replicates to be filtered.}

\item{violation}{Violation to be analyzed. Options are "homicidio", "secuestro",
"reclutamiento", and "desaparicion"}

\item{perp_change}{Change option: if there are victims in years after 2016 and
if the perpetrator "p_str" is "GUE-FARC", these victims become victims of
other guerrillas "GUE-OTRO"}
}
\value{
Data frame filter
}
\description{
Filter records to replicate results presented in the CEV methodology report.
}
\examples{
local_dir <- system.file("extdata", "right", package = "verdata")
replicates_data <- read_replicates(local_dir, "reclutamiento", 1, 2)
filter_standard_cev(replicates_data, "reclutamiento", perp_change = TRUE)

}
