% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prop_obs_rep.R
\name{prop_obs_rep}
\alias{prop_obs_rep}
\title{Function to obtain proportions of each
variable after to use combine_replicates}
\usage{
prop_obs_rep(data_obs_rep, strata_vars, na_obs = TRUE, digits = NULL)
}
\arguments{
\item{data_obs_rep}{a data frame containing the observed and
imputed values to be analyzed}

\item{strata_vars}{a vector of column names
identifying the variables to be used as strata in the analysis.}

\item{na_obs}{a logical value indicating whether or not to include missing
observations in the calculations (according to strata_vars).}

\item{digits}{Number of decimal places to round the proportion results to.}
}
\value{
Data frame that contains the proportions of each variable (from
summary_observed and combine_replicates)
}
\description{
Function to obtain proportions of each
variable after to use combine_replicates
}
\examples{
local_dir <- system.file("extdata", "right", package = "verdata")
replicates_data <- read_replicates(local_dir, "reclutamiento", 1, 2, "parquet")
tab_observed <- summary_observed("reclutamiento", replicates_data,
strata_vars = "sexo", strata_vars_com = "yy_hecho",
conflict_filter = TRUE, forced_dis = FALSE, edad_minors = TRUE)
tab_combine <- combine_replicates("reclutamiento", tab_observed,
replicates_data, strata_vars_rep = 'sexo', conflict_filter = TRUE,
forced_dis = FALSE, edad_minors = TRUE)
prop_data <- prop_obs_rep(tab_combine, "sexo", na_obs = TRUE, digits = 2)
}
