% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_replicates.R
\name{proportions_imputed}
\alias{proportions_imputed}
\title{proportions_imputed}
\usage{
proportions_imputed(complete_tab, strata_vars, digits = NULL)
}
\arguments{
\item{complete_tab}{A dataframe containing the output from \code{combine_replicates}.}

\item{strata_vars}{A vector of column names identifying the variables to be
used for stratification.}

\item{digits}{Number of decimal places to round the results to.}
}
\value{
A dataframe that contains the proportions after to apply
\code{combine_replicates}
}
\description{
Calculate the proportions of each level of a variable after
to calculate using \code{combine_replicates} on completed data (that includes
imputed values).
}
\examples{
local_dir <- system.file("extdata", "right", package = "verdata")
replicates_data <- read_replicates(local_dir, "reclutamiento", 1, 2, "parquet")
tab_observed <- summary_observed("reclutamiento", replicates_data,
strata_vars = "sexo", conflict_filter = FALSE, forced_dis_filter = FALSE, 
edad_minors_filter = FALSE, include_props = FALSE, prop_obs_na = FALSE, 
digits = 2)
tab_combine <- combine_replicates("reclutamiento", tab_observed,
replicates_data, strata_vars = 'sexo', conflict_filter = TRUE,
forced_dis_filter = FALSE, edad_minors_filter = FALSE)
prop_data_complete <- proportions_imputed(tab_combine, strata_vars = "sexo", 
digits = 2)
}
