% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratification.R
\docType{data}
\name{stratification}
\alias{stratification}
\title{Data documenting the stratifications used to replicate the results
of the methodological report of the joint JEP-CEV-HRDAG project
(version in English).}
\format{
A data frame with 31 rows and 4 variables
\describe{
\item{violation}{the human rights violation being analyzed}
\item{estimation}{the type of analysis the stratification was used for (e.g., patterns of violence by year, sex, etc.)}
\item{stratification}{the variables used to stratify the estimates}
\item{notes}{additional notes about the stratification; NA if no notes}
}
}
\source{
joint JEP-CEV-HRDAG project
}
\usage{
data(stratification)
}
\description{
Data documenting the stratifications used to replicate the results
of the methodological report of the joint JEP-CEV-HRDAG project
(version in English).
}
\keyword{datasets}
