% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_observed.R
\name{summary_observed}
\alias{summary_observed}
\title{Summary observed data}
\usage{
summary_observed(
  violation,
  data_rep,
  strata_vars_com = NULL,
  strata_vars = NULL,
  conflict_filter = TRUE,
  forced_dis = TRUE,
  edad_minors = TRUE,
  edad_na = FALSE
)
}
\arguments{
\item{violation}{Violation to be analyzed}

\item{data_rep}{Data frame containing the replicates}

\item{strata_vars_com}{Variable that will be analyzed and does not have
missing values (before imputation)}

\item{strata_vars}{Variable to be analyzed - before imputation
had missing values}

\item{conflict_filter}{Filter that indicates if the data is filtered by
the rule "is_conflict" or not.}

\item{forced_dis}{Filter that indicates if the data is filter by
the rule "is_forced_dis" or not.}

\item{edad_minors}{Optional filter by "edad" < 18.}

\item{edad_na}{Filter that indicates if the data is filtered by observations
(in "edad") that was not imputed (original in the data) = TRUE, or if shows
all observations (FALSE).}
}
\value{
Data frame with two or more columns, 1: name of variable(s) and 2;
number of observations in each category.
}
\description{
Summary observed data
}
\examples{
local_dir <- system.file("extdata", "right", package = "verdata")
replicates_data <- read_replicates(local_dir, "reclutamiento", 1, 2, "parquet")
tab_observed <- summary_observed("reclutamiento", replicates_data,
strata_vars = "sexo", strata_vars_com = "yy_hecho",
conflict_filter = TRUE, forced_dis = FALSE,
edad_minors = TRUE, edad_na = TRUE)
}
