% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_replicates.R
\name{combine_replicates}
\alias{combine_replicates}
\title{Combine replicates according to the Normal approximation using the laws of total expectation and variance.}
\usage{
combine_replicates(
  violation,
  data_obs,
  data_rep,
  strata_vars_rep,
  conflict_filter = TRUE,
  forced_dis = FALSE,
  edad_minors = FALSE
)
}
\arguments{
\item{violation}{Violation to be analyzed. Options are "homicidio", "secuestro",
"reclutamiento", and "desaparicion".}

\item{data_obs}{Dataframe that results from applying summary_observed.}

\item{data_rep}{Original data, not filtered by is_conflict.}

\item{strata_vars_rep}{Variable with all observations (without missing values).}

\item{conflict_filter}{Filter that indicates if the data is filtered using
the "is_conflict" rule.}

\item{forced_dis}{Filter that indicates if the data is filtered using the
"is_forced_dis" rule.}

\item{edad_minors}{Optional filter by age ("edad") < 18.}
}
\value{
A dataframe with 5 or more columns: name of variable(s), \code{observed}
the number of observations in each category for every variable, \code{imp_lo} the
lower bound of the 95\% confidence interval, \code{imp_hi} the upper bound of the
95\% confidence interval, and \code{imp_mean} the point estimate of the mean value.
}
\description{
Combine replicates according to the Normal approximation using the laws of total expectation and variance.
}
\examples{
local_dir <- system.file("extdata", "right", package = "verdata")
replicates_data <- read_replicates(local_dir, "reclutamiento", 1, 2, "parquet")
tab_observed <- summary_observed("reclutamiento", replicates_data,
strata_vars = "sexo", strata_vars_com = "yy_hecho",
conflict_filter = TRUE, is_conflict_na = TRUE, forced_dis = FALSE,
is_forced_dis_na = TRUE, edad_minors = TRUE, edad_na = FALSE, perp_na = FALSE,
sexo_na = FALSE, municipio_na = FALSE, etnia_na = FALSE)
tab_combine <- combine_replicates("reclutamiento", tab_observed,
replicates_data, strata_vars_rep = 'sexo', conflict_filter = TRUE,
forced_dis = FALSE, edad_minors = TRUE)
}
