% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_replicates.R
\name{combine_replicates}
\alias{combine_replicates}
\title{Combine replicates according to the Normal approximation using the laws of total expectation and variance.}
\usage{
combine_replicates(
  violation,
  replicates_obs_data,
  replicates_data,
  strata_vars,
  conflict_filter = TRUE,
  forced_dis_filter = FALSE,
  edad_minors_filter = FALSE,
  include_props = FALSE,
  digits = 2
)
}
\arguments{
\item{violation}{Violation to be analyzed. Options are "homicidio", "secuestro",
"reclutamiento" and "desaparicion".}

\item{replicates_obs_data}{Dataframe that results from applying \code{summary_observed}.}

\item{replicates_data}{Data frame containing replicates data.}

\item{strata_vars}{Variable with all observations (without missing values).}

\item{conflict_filter}{Filter that indicates if the data is filtered using
the "is_conflict" rule.}

\item{forced_dis_filter}{Filter that indicates if the data is filtered using the
"is_forced_dis" rule.}

\item{edad_minors_filter}{Optional filter by age ("edad") < 18.}

\item{include_props}{A logical value indicating whether or not to include
the proportions from the calculations before to merge with summary_observed's output.}

\item{digits}{Number of decimal places to round the results to (only relevant
if \code{include_props} is \code{TRUE}). Default value is 2.}
}
\value{
A dataframe with 5 or more columns: name of variable(s), \code{observed}
the number of observations in each category for every variable, \code{imp_lo} the
lower bound of the 95\% confidence interval, \code{imp_hi} the upper bound of the
95\% confidence interval, and \code{imp_mean} the point estimate of the mean value.
}
\description{
Combine replicates according to the Normal approximation using the laws of total expectation and variance.
}
\examples{
local_dir <- system.file("extdata", "right", package = "verdata")
replicates_data <- read_replicates(local_dir, "reclutamiento", 1, 2)
replicates_obs_data <- summary_observed("reclutamiento", replicates_data,
strata_vars = "sexo", conflict_filter = FALSE, forced_dis_filter = FALSE, 
edad_minors_filter = FALSE, include_props = FALSE, include_props_na = FALSE)
tab_combine <- combine_replicates("reclutamiento", replicates_obs_data,
replicates_data, strata_vars = 'sexo', conflict_filter = TRUE,
forced_dis_filter = FALSE, edad_minors_filter = FALSE, include_props = FALSE)
}
