% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_replicates.R
\name{combine_replicates}
\alias{combine_replicates}
\title{Combine replicates according to the Normal approximation using the laws of total expectation and variance}
\usage{
combine_replicates(
  violation,
  data_obs,
  data_rep,
  strata_vars_rep,
  conflict_filter = TRUE,
  forced_dis = FALSE,
  edad_minors = FALSE
)
}
\arguments{
\item{violation}{type of violation}

\item{data_obs}{Data frame that come from summary_observed's function}

\item{data_rep}{Original data, not filter for is_conflict.}

\item{strata_vars_rep}{variable with all observations (without missing values)}

\item{conflict_filter}{Filter that indicates if the data is filter for
the rule "is_conflict" or not.}

\item{forced_dis}{Filter that indicates if the data is filter for
the rule "is_forced_dis" or not.}

\item{edad_minors}{Optional filter by "edad" < 18.}
}
\value{
Data frame with 5 o more columns:
\enumerate{
\item Name of variable(s);
2.# Observations that shows each variable's category;
\item Confidence intervals (lower and upper) and median value imputed.
}
}
\description{
Combine replicates according to the Normal approximation using the laws of total expectation and variance
}
\examples{
local_dir <- system.file("extdata", "right", package = "verdata")
replicates_data <- read_replicates(local_dir, "reclutamiento", 1, 2, "parquet")
tab_observed <- summary_observed("reclutamiento", replicates_data,
strata_vars = "sexo", strata_vars_com = "yy_hecho",
conflict_filter = TRUE, is_conflict_na = TRUE, forced_dis = FALSE,
is_forced_dis_na = TRUE, edad_minors = TRUE, edad_na = FALSE, perp_na = FALSE,
sexo_na = FALSE, municipio_na = FALSE, etnia_na = FALSE)
tab_combine <- combine_replicates("reclutamiento", tab_observed,
replicates_data, strata_vars_rep = 'sexo', conflict_filter = TRUE,
forced_dis = FALSE, edad_minors = TRUE)
}
