% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confirm_files.R
\name{confirm_files}
\alias{confirm_files}
\title{Confirm files are identical to the ones published.}
\usage{
confirm_files(where_replicate, violacion, first_rep, last_rep)
}
\arguments{
\item{where_replicate}{File path for the folder containing the replicates.
The name of the files must include the violation in Spanish and lower case
letters (homicidio, secuestro, reclutamiento, desaparicion)}

\item{violacion}{Violation being analyzed. Options are "homicidio", "secuestro",
"reclutamiento", and "desaparicion".}

\item{first_rep}{First replicate in the range of replicates to be analyzed}

\item{last_rep}{Last replicate in the range of replicates to be analyzed.}
}
\value{
"You have the right file!" message if the files are identical to the
files published or error and "This file is not identical to the one published.
This means the results of the analysis may potentially be inconsistent." if
they are not.
}
\description{
Confirm files are identical to the ones published.
}
\examples{
local_dir <- system.file("extdata", "right", package = "verdata")
confirm_files(local_dir, "reclutamiento", 1, 2)
}
