% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_standard_cev.R
\name{filter_standard_cev}
\alias{filter_standard_cev}
\title{Filter to obtain the CEV's results - methodological document}
\usage{
filter_standard_cev(replicates, violation)
}
\arguments{
\item{replicates}{Data frame with the replicates}

\item{violation}{Violation to be analyzed}
}
\value{
Data frame filter
}
\description{
Filter to obtain the CEV's results - methodological document
}
\examples{
local_dir <- system.file("extdata", "right", package = "verdata")
replicates_data <- read_replicates(local_dir, "reclutamiento", 1, 2, "parquet")
filter_standard_cev(replicates_data, "reclutamiento")

}
